Replace uint64_t with uintptr_t for the data pointer in nvpair encoding. I feel certain I fixed this bug once before, but perhaps it changed upstream.  This was causing an untagged capability exception in ping(8).
(Debugged with the helpful eyes of David Chisnall.)

+++ sys/contrib/libnv/nvpair.c
@@ -113,7 +113,7 @@ nvpair_assert(const nvpair_t *nvp)
 }
 
 static nvpair_t *
-nvpair_allocv(const char *name, int type, uint64_t data, size_t datasize,
+nvpair_allocv(const char *name, int type, uintptr_t data, size_t datasize,
     size_t nitems)
 {
 	nvpair_t *nvp;
